# contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: ProcureIQ Sales API
  description: API for managing CRM Sales Orders, Products, Accounts, and Notifications in ProcureIQ.
  version: 1.2.0
paths:
...
  /api/v1/fieldservice/appointments:
    post:
      summary: Create service appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAppointmentRequest'
      responses:
        '201':
          description: Service appointment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceAppointmentResponse'
  /api/v1/fieldservice/appointments/{id}:
    get:
      summary: Retrieve service appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service appointment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceAppointmentResponse'
    put:
      summary: Update service appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAppointmentRequest'
      responses:
        '200':
          description: Service appointment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceAppointmentResponse'
    delete:
      summary: Soft-delete service appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Service appointment deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
  /api/v1/fieldservice/appointments/{appointmentId}/assign:
    post:
      summary: Assign service resource to service appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignResourceRequest'
      responses:
        '201':
          description: Service resource assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssignedResourceResponse'
  /api/v1/fieldservice/assigned-resources/{id}:
    delete:
      summary: Remove service resource assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Assignment removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
  /api/v1/fieldservice/appointments/{appointmentId}/candidates:
    get:
      summary: Retrieve candidate service resources for appointment using FSL constraints matching
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of candidate service resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceResourceResponse'
...
  /api/v1/fieldservice/asset-relationships:
    post:
      summary: Create asset relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRelationshipRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssetRelationshipResponse'
  /api/v1/fieldservice/asset-relationships/{id}:
    get:
      summary: Retrieve asset relationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssetRelationshipResponse'
    put:
      summary: Update asset relationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRelationshipRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssetRelationshipResponse'
    delete:
      summary: Delete asset relationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
  /api/v1/fieldservice/service-crews:
    post:
      summary: Create service crew
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCrewRequest'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceCrewResponse'
  /api/v1/fieldservice/service-crews/{id}:
    get:
      summary: Retrieve service crew
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceCrewResponse'
    put:
      summary: Update service crew
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCrewRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseServiceCrewResponse'
    delete:
      summary: Delete service crew
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
...
    ApiResponseString:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        data:
          type: string
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    ServiceResourceResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        resourceType:
          type: string
        isActive:
          type: boolean
...
    ServiceAppointmentRequest:
      type: object
      required:
        - parentId
        - parentType
      properties:
        parentId:
          type: integer
        parentType:
          type: string
          enum: [work_order, work_order_line_item]
        status:
          type: string
        earliestStartPermitted:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        arrivalWindowStart:
          type: string
          format: date-time
        arrivalWindowEnd:
          type: string
          format: date-time
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        duration:
          type: number
    ServiceAppointmentResponse:
      type: object
      properties:
        id:
          type: integer
        parentId:
          type: integer
        parentType:
          type: string
        status:
          type: string
        earliestStartPermitted:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        arrivalWindowStart:
          type: string
          format: date-time
        arrivalWindowEnd:
          type: string
          format: date-time
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        duration:
          type: number
    ApiResponseServiceAppointmentResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceAppointmentResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
    AssignResourceRequest:
      type: object
      required:
        - serviceResourceId
      properties:
        serviceResourceId:
          type: integer
    AssignedResourceResponse:
      type: object
      properties:
        id:
          type: integer
        serviceAppointmentId:
          type: integer
        serviceResourceId:
          type: integer
    ApiResponseAssignedResourceResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/AssignedResourceResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...
    AssetRelationshipRequest:
      type: object
      required:
        - assetId
        - relatedAssetId
        - relationshipType
      properties:
        assetId:
          type: integer
        relatedAssetId:
          type: integer
        relationshipType:
          type: string
    AssetRelationshipResponse:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        relatedAssetId:
          type: integer
        relationshipType:
          type: string
    ApiResponseAssetRelationshipResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/AssetRelationshipResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
    ServiceCrewRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    ServiceCrewResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ApiResponseServiceCrewResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 201
        data:
          $ref: '#/components/schemas/ServiceCrewResponse'
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
...